Squashed 'libraries/textogl/' changes from 562a0eb94..7ecab7e23
7ecab7e23 don't use internal glm types
cfb5eb462 fix warnings
9bec17e89 remove newer cmake features to actually support v3.0

git-subtree-dir: libraries/textogl
git-subtree-split: 7ecab7e23c49840b75385cb2f7402045fc4856b3

+++ demo/demo.cpp
@@ -111,7 +111,7 @@ int main(int argc, char * argv[])
 
         for(std::size_t i = 0; i < font_arr.size(); ++i)
             static_arr[i].render_text(textogl::Color{0.0f, 0.0f, 0.0f, 1.0f}, textogl::Vec2<float>{(float)win.getSize().x, (float)win.getSize().y},
-                textogl::Vec2<float>{win.getSize().x, i * 60}, textogl::ORIGIN_VERT_TOP | textogl::ORIGIN_HORIZ_RIGHT);
+                textogl::Vec2<float>{(float)win.getSize().x, (float)(i * 60)}, textogl::ORIGIN_VERT_TOP | textogl::ORIGIN_HORIZ_RIGHT);
 
         win.display();
     }
